Add useStatusBar hook and update views
This commit adds a new custom hook called useStatusBar, which is used to set the background color of the status bar. It is imported and used in the Home, Chart, Stock, Trade, and Profile views. Additionally, the styleSheet file is updated to include utility functions for calculating viewport units and defining common styles.